Moves all middleware to `Middleware` subnamespace
We have a lot of middleware at this point, so they need to be moved to a
subnamespace.
